encapsulate most of gp global variables in GP_DATA. move gp output to es.c define TIMERstart()

+++ src/gp/gp.h
@@ -25,7 +25,7 @@ int term_height(void);
 int term_width(void);
 void hit_return(void);
 
-extern ulong secure;
+extern int secure;
 extern ulong init_opts;
 extern char *current_logfile;
 extern jmp_buf environnement;
